pyproject.toml
[tool.poetry]
name = "freedom-property-hubspot"
version = "0.1.0"
description = ""
authors = ["HM <hieumdd@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.9"
requests = "^2.26.0"
google-cloud-bigquery = "^2.31.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
mypy = "^0.910"
types-requests = "^2.26.1"
pytest = "^6.2.5"
pytest-timeout = "^2.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-p no:warnings -s --durations=0"
testpaths = [
    "test",
]
timeout = 530
